Move Login and Register to auth screen
