+++ src/components/index.js
@@ -1,2 +1,2 @@
-export * from './Registration';
-export * from './Login';
+export * from '../screens/auth/Registration';
+export * from '../screens/auth/Login';
